S2D-55 refactor : Comment관련 Controller, Service 분리
