+++ src/main/java/com/sluv/server/domain/comment/service/CommentService.java
@@ -1,31 +1,16 @@
 package com.sluv.server.domain.comment.service;
 
-import com.sluv.server.domain.closet.entity.Closet;
-import com.sluv.server.domain.closet.repository.ClosetRepository;
-import com.sluv.server.domain.comment.dto.CommentImgDto;
-import com.sluv.server.domain.comment.dto.CommentItemResDto;
 import com.sluv.server.domain.comment.dto.CommentPostReqDto;
-import com.sluv.server.domain.comment.dto.CommentReportPostReqDto;
 import com.sluv.server.domain.comment.dto.CommentResDto;
 import com.sluv.server.domain.comment.dto.SubCommentPageResDto;
 import com.sluv.server.domain.comment.entity.Comment;
-import com.sluv.server.domain.comment.entity.CommentImg;
-import com.sluv.server.domain.comment.entity.CommentItem;
 import com.sluv.server.domain.comment.entity.CommentLike;
-import com.sluv.server.domain.comment.entity.CommentReport;
 import com.sluv.server.domain.comment.exception.CommentNotFoundException;
-import com.sluv.server.domain.comment.exception.CommentReportDuplicateException;
-import com.sluv.server.domain.comment.repository.CommentImgRepository;
-import com.sluv.server.domain.comment.repository.CommentItemRepository;
+import com.sluv.server.domain.comment.handler.CommentHelper;
+import com.sluv.server.domain.comment.manager.CommentImgManager;
+import com.sluv.server.domain.comment.manager.CommentItemManager;
 import com.sluv.server.domain.comment.repository.CommentLikeRepository;
-import com.sluv.server.domain.comment.repository.CommentReportRepository;
 import com.sluv.server.domain.comment.repository.CommentRepository;
-import com.sluv.server.domain.item.entity.Item;
-import com.sluv.server.domain.item.entity.ItemImg;
-import com.sluv.server.domain.item.exception.ItemNotFoundException;
-import com.sluv.server.domain.item.repository.ItemImgRepository;
-import com.sluv.server.domain.item.repository.ItemRepository;
-import com.sluv.server.domain.item.repository.ItemScrapRepository;
 import com.sluv.server.domain.question.entity.Question;
 import com.sluv.server.domain.question.exception.QuestionNotFoundException;
 import com.sluv.server.domain.question.repository.QuestionRepository;
@@ -45,18 +30,53 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j
 @RequiredArgsConstructor
 public class CommentService {
-    private final CommentRepository commentRepository;
-    private final CommentImgRepository commentImgRepository;
-    private final CommentItemRepository commentItemRepository;
-    private final CommentLikeRepository commentLikeRepository;
-    private final CommentReportRepository commentReportRepository;
     private final QuestionRepository questionRepository;
-    private final ItemRepository itemRepository;
-    private final ItemImgRepository itemImgRepository;
-    private final ItemScrapRepository itemScrapRepository;
-    private final ClosetRepository closetRepository;
+    private final CommentRepository commentRepository;
+    private final CommentLikeRepository commentLikeRepository;
     private final AiModelService aiModelService;
+    private final CommentHelper commentHelper;
+    private final CommentItemManager commentItemManager;
+    private final CommentImgManager commentImgManager;
 
+    /**
+     * 댓글 조회
+     */
+    @Transactional(readOnly = true)
+    public PaginationResDto<CommentResDto> getComment(User user, Long questionId, Pageable pageable) {
+        // 해당 페이지 검색
+        Page<Comment> commentPage = commentRepository.getAllQuestionComment(questionId, pageable);
+
+        // Content 제작
+        List<CommentResDto> content = commentHelper.getCommentResDtos(user, commentPage);
+
+        return PaginationResDto.of(commentPage, content);
+    }
+
+    /**
+     * 대댓글 조회
+     */
+    @Transactional(readOnly = true)
+    public SubCommentPageResDto<CommentResDto> getSubComment(User user, Long commentId, Pageable pageable) {
+        // 대댓글 페이지 검색
+        Page<Comment> commentPage = commentRepository.getAllSubComment(commentId, pageable);
+
+        // Content 제작
+        List<CommentResDto> content = commentHelper.getCommentResDtos(user, commentPage);
+
+        // 남은 댓글 수. 총 댓글 수 - ((현재 페이지 +1)*페이지당 size)가 0보다 작으면 0, 아닐 경우 해당 값
+        long restCommentNum =
+                commentPage.getTotalElements() - ((long) (commentPage.getNumber() + 1) * commentPage.getSize()) >= 0
+                        ? commentPage.getTotalElements() - ((long) (commentPage.getNumber() + 1)
+                        * commentPage.getSize())
+                        : 0;
+
+        return SubCommentPageResDto.of(commentPage, content, restCommentNum);
+
+    }
+
+    /**
+     * 댓글 등록
+     */
     @Transactional
     public void postComment(User user, Long questionId, CommentPostReqDto dto) {
         log.info("댓글 등록 - 사용자 : {}, 질문 게시글 : {}, 댓글 내용 {}", user.getId(), questionId, dto.getContent());
@@ -66,11 +86,15 @@ public class CommentService {
                 Comment.toEntity(user, question, dto.getContent())
         );
 
-        saveCommentImgAndItem(dto, comment);
+        commentItemManager.saveCommentItem(dto, comment);
+        commentImgManager.saveCommentImg(dto, comment);
         aiModelService.censorComment(comment);
 
     }
 
+    /**
+     * 대댓글 등록
+     */
     @Transactional
     public void postSubComment(User user, Long questionId, Long commentId, CommentPostReqDto dto) {
         log.info("대댓글 등록 - 사용자 : {}, 질문 게시글 : {}, 상위 댓글 : {}, 댓글 내용 {}",
@@ -85,56 +109,14 @@ public class CommentService {
                 Comment.toEntity(user, question, dto.getContent(), parentComment)
         );
 
-        saveCommentImgAndItem(dto, comment);
+        commentItemManager.saveCommentItem(dto, comment);
+        commentImgManager.saveCommentImg(dto, comment);
         aiModelService.censorComment(comment);
     }
 
     /**
-     * 댓글에 이미지 및 아이템 등록. Dto에 이미지 및 아이템이 존재한다면 등록 메소드 호출.
+     * 댓글 수정
      */
-    private void saveCommentImgAndItem(CommentPostReqDto dto, Comment comment) {
-        // 1. CommentImg 등록
-        if (dto.getImgList() != null) {
-            saveCommentImg(dto, comment);
-        }
-
-        // 2. CommentItem 등록
-        if (dto.getItemList() != null) {
-            saveCommentItem(dto, comment);
-        }
-    }
-
-    private void saveCommentItem(CommentPostReqDto dto, Comment comment) {
-        // 초기화
-        commentItemRepository.deleteAllByCommentId(comment.getId());
-        if (dto.getItemList() != null) {
-            // dto로 부터 새로운 CommentItem 생성
-            List<CommentItem> itemList = dto.getItemList().stream().map(itemReqDto -> {
-                        Item item = itemRepository.findById(itemReqDto.getItemId()).orElseThrow(ItemNotFoundException::new);
-                        return CommentItem.toEntity(comment, item, itemReqDto);
-                    }
-            ).toList();
-
-            // 저장
-            commentItemRepository.saveAll(itemList);
-        }
-    }
-
-    private void saveCommentImg(CommentPostReqDto dto, Comment comment) {
-        // 초기화
-        commentImgRepository.deleteAllByCommentId(comment.getId());
-
-        // dto로 부터 새로운 CommentImg 생성
-        if (dto.getImgList() != null) {
-            List<CommentImg> imgList = dto.getImgList().stream().map(imgUrl ->
-                    CommentImg.toEntity(comment, imgUrl)
-            ).toList();
-
-            // 저장
-            commentImgRepository.saveAll(imgList);
-        }
-    }
-
     @Transactional
     public void putComment(User user, Long commentId, CommentPostReqDto dto) {
         /**
@@ -151,41 +133,15 @@ public class CommentService {
         // content 변경
         comment.changeContent(dto.getContent());
 
-        saveCommentImgAndItem(dto, comment);
+        commentItemManager.saveCommentItem(dto, comment);
+        commentImgManager.saveCommentImg(dto, comment);
         aiModelService.censorComment(comment);
 
     }
 
-    @Transactional
-    public void postCommentLike(User user, Long commentId) {
-        log.info("댓글 삭제 - 사용자 : {}, 댓글 : {}", user.getId(), commentId);
-        Boolean commentListStatus = commentLikeRepository.existsByUserIdAndCommentId(user.getId(), commentId);
-
-        if (!commentListStatus) {
-            Comment comment = commentRepository.findById(commentId).orElseThrow(CommentNotFoundException::new);
-            commentLikeRepository.save(
-                    CommentLike.toEntity(user, comment)
-            );
-        } else {
-            commentLikeRepository.deleteByUserIdAndCommentId(user.getId(), commentId);
-        }
-    }
-
-    @Transactional
-    public void postCommentReport(User user, Long commentId, CommentReportPostReqDto dto) {
-        log.info("댓글 신고 - 사용자 : {}, 신고 댓글 : {}", user.getId(), commentId);
-        Boolean reportStatus = commentReportRepository.existsByReporterIdAndCommentId(user.getId(), commentId);
-
-        if (reportStatus) {
-            throw new CommentReportDuplicateException();
-        }
-
-        Comment comment = commentRepository.findById(commentId).orElseThrow(CommentNotFoundException::new);
-        commentReportRepository.save(
-                CommentReport.toEntity(comment, user, dto)
-        );
-    }
-
+    /**
+     * 댓글 삭제
+     */
     @Transactional
     public void deleteComment(User user, Long commentId) {
         log.info("댓글 삭제 - 사용자 : {}, 댓글 : {}", user.getId(), commentId);
@@ -204,72 +160,22 @@ public class CommentService {
 
     }
 
-    @Transactional(readOnly = true)
-    public PaginationResDto<CommentResDto> getComment(User user, Long questionId, Pageable pageable) {
-        // 해당 페이지 검색
-        Page<Comment> commentPage = commentRepository.getAllQuestionComment(questionId, pageable);
-
-        // Content 제작
-        List<CommentResDto> content = getCommentResDtos(user, commentPage);
-
-        return PaginationResDto.of(commentPage, content);
-    }
-
-    @Transactional(readOnly = true)
-    public SubCommentPageResDto<CommentResDto> getSubComment(User user, Long commentId, Pageable pageable) {
-        // 대댓글 페이지 검색
-        Page<Comment> commentPage = commentRepository.getAllSubComment(commentId, pageable);
-
-        // Content 제작
-        List<CommentResDto> content = getCommentResDtos(user, commentPage);
-
-        // 남은 댓글 수. 총 댓글 수 - ((현재 페이지 +1)*페이지당 size)가 0보다 작으면 0, 아닐 경우 해당 값
-        long restCommentNum =
-                commentPage.getTotalElements() - ((long) (commentPage.getNumber() + 1) * commentPage.getSize()) >= 0
-                        ? commentPage.getTotalElements() - ((long) (commentPage.getNumber() + 1)
-                        * commentPage.getSize())
-                        : 0;
-
-        return SubCommentPageResDto.of(commentPage, content, restCommentNum);
-
-    }
-
-    private List<CommentResDto> getCommentResDtos(User user, Page<Comment> commentPage) {
-        return commentPage
-                .stream()
-                .map(comment -> {
-
-                    // 해당 Comment에 해당하는 이미지 조회
-                    List<CommentImgDto> imgList = commentImgRepository.findAllByCommentId(comment.getId())
-                            .stream().map(CommentImgDto::of)
-                            .toList();
-                    // 해당 Comment에 해당하는 아이템 조회
-                    List<CommentItemResDto> itemList = commentItemRepository.findAllByCommentId(comment.getId())
-                            .stream().map(commentItem -> getCommentItemResDto(commentItem, user))
-                            .toList();
-
-                    // 해당 Comment의 좋아요 수
-                    Integer likeNum = commentLikeRepository.countByCommentId(comment.getId());
-                    // 현재 유저의 해당 Comment 좋아요 여부
-                    Boolean likeStatus = commentLikeRepository.existsByUserIdAndCommentId(user.getId(),
-                            comment.getId());
-
-                    return CommentResDto.of(comment, user, imgList, itemList, likeNum, likeStatus);
-                }).toList();
-    }
-
     /**
-     * CommentItem -> CommentItemResDto 변경하는 메소드
-     *
-     * @param commentItem
-     * @return CommentItemResDto
+     * 댓글 좋아요
      */
-    private CommentItemResDto getCommentItemResDto(CommentItem commentItem, User user) {
-        ItemImg mainImg = itemImgRepository.findMainImg(commentItem.getItem().getId());
-        List<Closet> closetList = closetRepository.findAllByUserId(user.getId());
-        Boolean itemScrapStatus = itemScrapRepository.getItemScrapStatus(commentItem.getItem(), closetList);
+    @Transactional
+    public void postCommentLike(User user, Long commentId) {
+        log.info("댓글 삭제 - 사용자 : {}, 댓글 : {}", user.getId(), commentId);
+        Boolean commentListStatus = commentLikeRepository.existsByUserIdAndCommentId(user.getId(), commentId);
 
-        return CommentItemResDto.of(commentItem, mainImg, itemScrapStatus);
+        if (!commentListStatus) {
+            Comment comment = commentRepository.findById(commentId).orElseThrow(CommentNotFoundException::new);
+            commentLikeRepository.save(
+                    CommentLike.toEntity(user, comment)
+            );
+        } else {
+            commentLikeRepository.deleteByUserIdAndCommentId(user.getId(), commentId);
+        }
     }
